pyproject.toml
[project]
name = "district-tool"
version = "0.0.1"
description = "Web-based design-exploration tool for environmental and energy performance of European urban blocks"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "John Diamant-Boustead", email = "john@example.org"},
]
license = {text = "Proprietary"}

dependencies = [
    # Spatial core
    "geopandas>=1.0",
    "shapely>=2.0",
    "rasterio>=1.3",
    "pyproj>=3.6",
    "fiona>=1.9",
    "pyogrio>=0.7",
    "numpy>=1.26",
    # Catalogue
    "pyyaml>=6.0",
    "requests>=2.31",
    # Configuration
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "ruff>=0.3",
    "mypy>=1.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B"]
ignore = ["E501"]  # line length handled by formatter

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"

[tool.mypy]
python_version = "3.11"
strict = false
warn_unused_ignores = true
warn_return_any = true
